// src/types/data.ts
export type BackendDepartmentDataItem = {
  Дата: string;
  Департамент: string;
  "Кол-во отработанных часов": string;
  Команда: string;
  Отдел: string;
};

export type DepartmentDataItem = {
  department: string;
  division: string;
  team: string;
  month: string;
  date: string;
  workHours: string;
};